+++ audioware/src/types/mod.rs
@@ -1,3 +1,4 @@
 pub mod bank;
+pub mod id;
 pub mod redmod;
 pub mod voice;
